Replace `MaybeCompressor<T>` construct with `NO_COMPRESSION` constant
