+++ bincache/src/strategies.rs
@@ -1,11 +1,7 @@
 mod disk;
 mod hybrid;
 mod memory;
-mod noop;
 
 pub use disk::Disk;
 pub use hybrid::{Hybrid, Limits};
 pub use memory::Memory;
-
-#[cfg(test)]
-pub(crate) use noop::Noop;
