+++ bincache/src/strategies/disk.rs
@@ -179,12 +179,12 @@ impl RecoverableStrategy for Disk {
 #[cfg(test)]
 mod tests {
     use super::{Disk, LIMIT_KIND_BYTE, LIMIT_KIND_ENTRY};
-    use crate::{async_test, compression::MaybeCompressor, utils::test::TempDir, Cache, Error};
+    use crate::{async_test, utils::test::TempDir, Cache, Error, NO_COMPRESSION};
 
     async_test! {
         async fn test_default() {
             let temp_dir = TempDir::new();
-            let mut cache = Cache::new(Disk::new(temp_dir.as_ref(), None, None), MaybeCompressor::noop());
+            let mut cache = Cache::new(Disk::new(temp_dir.as_ref(), None, None), NO_COMPRESSION);
 
             cache.put("foo", b"foo".to_vec()).await.unwrap();
 
@@ -214,7 +214,7 @@ mod tests {
 
         async fn test_strategy_with_byte_limit() {
             let temp_dir = TempDir::new();
-            let mut cache = Cache::new(Disk::new(temp_dir.as_ref(), Some(6), None), MaybeCompressor::noop());
+            let mut cache = Cache::new(Disk::new(temp_dir.as_ref(), Some(6), None), NO_COMPRESSION);
 
             let foo_data = b"foo".to_vec();
             let bar_data = b"bar".to_vec();
@@ -243,7 +243,7 @@ mod tests {
 
         async fn test_strategy_with_entry_limit() {
             let temp_dir = TempDir::new();
-            let mut cache = Cache::new(Disk::new(temp_dir.as_ref(), None, Some(3)), MaybeCompressor::noop());
+            let mut cache = Cache::new(Disk::new(temp_dir.as_ref(), None, Some(3)), NO_COMPRESSION);
 
             cache.put("foo", b"foo".to_vec()).await.unwrap();
             cache.put("bar", b"bar".to_vec()).await.unwrap();
@@ -267,7 +267,7 @@ mod tests {
 
             // populate cache
             {
-                let mut cache = Cache::new(Disk::new(temp_dir.as_ref(), None, None), MaybeCompressor::noop());
+                let mut cache = Cache::new(Disk::new(temp_dir.as_ref(), None, None), NO_COMPRESSION);
 
                 cache.put("foo", b"foo".to_vec()).await.unwrap();
                 cache.put("bar", b"bar".to_vec()).await.unwrap();
@@ -275,7 +275,7 @@ mod tests {
 
             // recover cache
             {
-                let mut cache = Cache::new(Disk::new(temp_dir.as_ref(), None, None), MaybeCompressor::noop());
+                let mut cache = Cache::new(Disk::new(temp_dir.as_ref(), None, None), NO_COMPRESSION);
                 let recovered_items = cache
                     .recover(|k| Some(k.to_string()))
                     .await
