+++ bincache/src/strategies/memory.rs
@@ -1,10 +1,7 @@
 use async_trait::async_trait;
 use std::borrow::Cow;
 
-use crate::{
-    traits::{CacheKey, CacheStrategy},
-    Result,
-};
+use crate::{CacheKey, CacheStrategy, Result};
 
 const LIMIT_KIND_BYTE: &str = "Stored bytes";
 const LIMIT_KIND_ENTRY: &str = "Stored entries";
@@ -102,11 +99,11 @@ impl CacheStrategy for Memory {
 #[cfg(test)]
 mod tests {
     use super::{Memory, LIMIT_KIND_BYTE, LIMIT_KIND_ENTRY};
-    use crate::{async_test, compression::MaybeCompressor, Cache, Error};
+    use crate::{async_test, Cache, Error, NO_COMPRESSION};
 
     async_test! {
         async fn test_default_strategy() {
-            let mut cache = Cache::new(Memory::default(), MaybeCompressor::noop());
+            let mut cache = Cache::new(Memory::default(), NO_COMPRESSION);
 
             cache.put("foo", b"foo".to_vec()).await.unwrap();
 
@@ -135,7 +132,7 @@ mod tests {
         }
 
         async fn test_strategy_with_byte_limit() {
-            let mut cache = Cache::new(Memory::new(Some(6), None), MaybeCompressor::noop());
+            let mut cache = Cache::new(Memory::new(Some(6), None), NO_COMPRESSION);
 
             cache.put("foo", b"foo".to_vec()).await.unwrap();
             cache.put("bar", b"bar".to_vec()).await.unwrap();
@@ -154,7 +151,7 @@ mod tests {
         }
 
         async fn test_strategy_with_entry_limit() {
-            let mut cache = Cache::new(Memory::new(None, Some(3)), MaybeCompressor::noop());
+            let mut cache = Cache::new(Memory::new(None, Some(3)), NO_COMPRESSION);
 
             cache.put("foo", b"foo".to_vec()).await.unwrap();
             cache.put("bar", b"bar".to_vec()).await.unwrap();
